refactor: main package for ice
