+++ examples/basic-project/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@ice/app": "file:../../packages/ice",
     "@ice/runtime": "^1.0.0",
+    "@ice/plugin-auth": "^1.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
