+++ packages/ice/package.json
@@ -21,7 +21,6 @@
   "dependencies": {
     "@builder/pack": "^0.6.0",
     "@ice/webpack-config": "^1.0.0",
-    "@ice/plugin-app": "^1.0.0",
     "@ice/plugin-auth": "^1.0.0",
     "@ice/route-manifest": "^1.0.0",
     "@ice/runtime": "^1.0.0",
