--- packages/ice/src/constant.ts
@@ -1,4 +1,0 @@
-export const builtInPlugins = [
-  '@ice/plugin-app',
-  '@ice/plugin-auth',
-];
