+++ packages/ice/src/tasks/ssr/serverRender.ts
@@ -1,5 +1,6 @@
 import * as fs from 'fs';
-import { matchRoutes } from 'react-router-dom';
+import { matchRoutes } from '@ice/runtime';
+import type { Request, Response } from 'express';
 
 interface Options {
   routeManifest: string;
@@ -12,13 +13,13 @@ export function setupRenderServer(options: Options) {
     serverCompiler,
   } = options;
 
-  return async (req, res) => {
+  return async (req: Request, res: Response) => {
     // Read the latest routes info.
     const routes = JSON.parse(fs.readFileSync(routeManifest, 'utf8'));
 
     // If not match pages routes, hand over to webpack dev server for processing
     let matches = matchRoutes(routes, req.path);
-    if (!matches) return;
+    if (matches.length === 0) return;
 
     const entry = await serverCompiler();
     const serverEntry = await import(entry);
