+++ packages/ice/src/utils/getContextConfig.ts
@@ -1,12 +1,9 @@
-import * as path from 'path';
-import { createRequire } from 'module';
 import type { Context } from 'build-scripts';
 import type { Config } from '@ice/types';
 import type { Configuration } from 'webpack';
 import type { Configuration as DevServerConfiguration } from 'webpack-dev-server';
 import type { UnpluginOptions } from 'unplugin';
 import { getWebpackConfig, getTransformPlugins as getBuiltInPlugins } from '@ice/webpack-config';
-import { builtInPlugins } from '../constant.js';
 
 export interface ContextConfig {
   name: string;
@@ -16,7 +13,6 @@ export interface ContextConfig {
   };
   getTransformPlugins: (config?: Partial<Config>) => UnpluginOptions[];
 }
-const require = createRequire(import.meta.url);
 
 function getContextConfig(context: Context<Config>): ContextConfig[] {
   const { getConfig, rootDir } = context;
@@ -26,19 +22,9 @@ function getContextConfig(context: Context<Config>): ContextConfig[] {
   }
   const configs = contextConfig.map(({ config, name }) => {
     // add runtime alias for built-in plugins
-    const builtInAlias = {};
-    builtInPlugins.forEach((pluginName) => {
-      builtInAlias[`${pluginName}/runtime`] = path.join(require.resolve(pluginName), '../../runtime');
-    });
     const webpackConfig = getWebpackConfig({
       rootDir,
-      config: {
-        ...config,
-        alias: {
-          ...(config.alias || {}),
-          ...builtInAlias,
-        },
-      },
+      config,
     });
     const getTransformPlugins = (customConfig?: Partial<Config>) => {
       return getBuiltInPlugins(rootDir, {
