--- packages/plugin-app/CHANGELOG.md
@@ -1,6 +1,0 @@
-# Changelog
-
-## v1.0.0
-
-- [feat] register basic webpack task
-- [feat] support render `src/Document` when start dev server
