--- packages/plugin-app/README.md
@@ -1,4 +1,0 @@
-# @ice/plugin-app
-
-Built-In plugin for web framework `ice`.
-This package include basic webpack task for `ice`.
