--- packages/plugin-app/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@ice/plugin-app",
-  "version": "1.0.0",
-  "description": "plugin app",
-  "type": "module",
-  "main": "./esm/index.js",
-  "exports": {
-    ".": "./esm/index.js",
-    "./runtime": "./runtime/index.ts"
-  },
-  "author": "ice-admin",
-  "files": [
-    "runtime",
-    "esm"
-  ],
-  "license": "MIT",
-  "repository": "ice-lab/ice-next",
-  "bugs": "https://github.com/ice-lab/ice-next/issues",
-  "homepage": "https://next.ice.work",
-  "dependencies": {
-    "chalk": "^4.0.0",
-    "cross-spawn": "^7.0.3",
-    "open": "^8.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "unplugin": "^0.3.2"
-  },
-  "devDependencies": {
-    "@ice/types": "^1.0.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "react-router-dom": "^6.2.2"
-  }
-}
