--- packages/plugin-app/src/userConfig.ts
@@ -1,113 +1,0 @@
-import chalk from 'chalk';
-import type { Config } from '@ice/types';
-import type { Context } from 'build-scripts';
-
-const mergeDefaultValue = <T>(config: Config, key: string, value: T): Config => {
-  if (value) {
-    if (typeof value === 'object') {
-      return {
-        ...config,
-        [key]: {
-          ...(config[key] || {}),
-          ...value,
-        },
-      };
-    } else {
-      config[key] = value;
-      return config;
-    }
-  }
-  return config;
-};
-
-const userConfig = [
-  {
-    name: 'alias',
-    validation: 'object',
-    setConfig: (config: Config, alias: Config['alias']) => {
-      return mergeDefaultValue(config, 'alias', alias);
-    },
-  },
-  {
-    name: 'define',
-    validation: 'object',
-    setConfig: (config: Config, define: Config['define']) => {
-      return mergeDefaultValue(config, 'define', define);
-    },
-  },
-  {
-    name: 'devPublicPath',
-    validation: 'string',
-    setConfig: (config: Config, publicPath: Config['publicPath'], context: Context) => {
-      return mergeDefaultValue(config, 'publicPath', context.command === 'start' && publicPath);
-    },
-  },
-  {
-    name: 'publicPath',
-    validation: 'string',
-    setConfig: (config: Config, publicPath: Config['publicPath'], context: Context) => {
-      return mergeDefaultValue(config, 'publicPath', context.command === 'build' && publicPath);
-    },
-  },
-  {
-    name: 'hash',
-    validation: 'string|boolean',
-    setConfig: (config: Config, hash: Config['hash']) => {
-      return mergeDefaultValue(config, 'hash', hash);
-    },
-  },
-  /* {
-    name: 'minify',
-    validation: 'boolean|string',
-  }, */
-  {
-    name: 'externals',
-    validation: 'object',
-    setConfig: (config: Config, externals: Config['externals']) => {
-      return mergeDefaultValue(config, 'externals', externals);
-    },
-  },
-  {
-    name: 'outputDir',
-    validation: 'string',
-    setConfig: (config: Config, outputDir: Config['outputDir']) => {
-      return mergeDefaultValue(config, 'outputDir', outputDir);
-    },
-  },
-  {
-    name: 'proxy',
-    validation: 'object',
-    setConfig: (config: Config, proxy: Config['proxy']) => {
-      return mergeDefaultValue(config, 'proxy', proxy);
-    },
-  },
-  {
-    name: 'filename',
-    validation: 'string',
-    setConfig: (config: Config, filename: string) => {
-      if (filename) {
-        const configFilename: Config['configureWebpack'][0] = (webpackConfig) => {
-          webpackConfig.output.filename = filename;
-          return webpackConfig;
-        };
-        config.configureWebpack = [...(config.configureWebpack || []), configFilename];
-      }
-    },
-  },
-  {
-    name: 'webpack',
-    validation: 'function',
-    setConfig: (config: Config, configureWebpack: Config['configureWebpack'][0]) => {
-      if (configureWebpack) {
-        // create warning for user
-        const customConfigWebpack: Config['configureWebpack'][0] = (...args) => {
-          console.log(chalk.yellowBright('[WARN]', 'It is not recommended to configure webpack directly.'));
-          return configureWebpack(...args);
-        };
-        config.configureWebpack = [...(config.configureWebpack || []), customConfigWebpack];
-      }
-    },
-  },
-];
-
-export default userConfig;
