--- packages/plugin-app/tests/index.test.ts
@@ -1,7 +1,0 @@
-import ice from '../src/index';
-
-describe('basic test', () => {
-  it('test return value', () => {
-    expect(true).toBe(true);
-  });
-});
