--- packages/plugin-app/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": "./",
-    "rootDir": "src",
-    "outDir": "esm"
-  },
-  "include": ["src"]
-}
