+++ packages/plugin-auth/src/index.ts
@@ -1,10 +1,15 @@
+import * as path from 'path';
+import { fileURLToPath } from 'url';
 import type { Plugin } from '@ice/types';
 
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
 const plugin: Plugin = ({ generator }) => {
+  const runtimePath = path.join(__dirname, '../runtime');
   // 注册 API：import { useAuth, withAuth } from 'ice';
   generator.addExport({
     specifier: ['withAuth', 'useAuth'],
-    source: '@ice/plugin-auth/runtime/Auth',
+    source: path.join(runtimePath, 'Auth'),
   });
 
   // 注册类型：appConfig.auth
@@ -13,7 +18,7 @@ const plugin: Plugin = ({ generator }) => {
   // }
   generator.addConfigTypes({
     specifier: ['AuthConfig'],
-    source: '@ice/plugin-auth/runtime/types',
+    source: path.join(runtimePath, 'types'),
     type: true,
     exportAlias: {
       AuthConfig: 'auth?',
