+++ packages/types/src/userConfig.ts
@@ -1,3 +1,4 @@
+import type { IPluginList } from 'build-scripts';
 import type { Config, ModifyWebpackConfig } from './config';
 
 export interface UserConfig {
@@ -11,4 +12,5 @@ export interface UserConfig {
   proxy?: Config['proxy'];
   filename?: string;
   webpack?: ModifyWebpackConfig;
+  plugins?: IPluginList;
 }
