+++ tsconfig.json
@@ -5,7 +5,6 @@
     { "path": "packages/route-manifest" },
     { "path": "packages/types" },
     { "path": "packages/build-webpack-config" },
-    { "path": "packages/plugin-app"},
     { "path": "packages/ice" },
     { "path": "packages/plugin-auth" }
   ]
